Use tpolecat plugin

+++ src/test/scala/io/apibuilder/validation/helpers/PerformanceHelpers.scala
@@ -5,7 +5,7 @@ trait PerformanceHelpers {
   def time(numberIterations: Int)(f: Long => Any): Long = {
     val start = System.currentTimeMillis()
     0.to(numberIterations).foreach { i =>
-      f(i)
+      f(i.toLong)
     }
     System.currentTimeMillis() - start
   }
